fix some imports and add toogle playing function
